requirements.txt
sanic==22.6.2
transformers
accelerate
timm
numpy
torch
opencv-python-headless
Pillow
ipython
torchvision